Refactor back arrow, fix keyDown on editing deck
