+++ src/components/BackArrow/BackArrow.module.css
@@ -1,0 +1,24 @@
+.topBar {
+    display: flex;
+    flex-direction: row;
+    padding-bottom: 20px;
+}
+
+.cardSettingsIcon {
+    opacity: 0.4;
+    transition: 0.4s;
+}
+
+.cardSettingsIcon:hover {
+    opacity: 1.0;
+    cursor: pointer;
+}
+
+.emptyBackArrow {
+    width: 1.25rem;
+    height: 2rem;
+}
+
+.navigation {
+    padding-left: 4%;
+}
