+++ src/components/BackArrow/index.ts
@@ -1,0 +1,3 @@
+import BackArrow from './BackArrow';
+
+export default BackArrow;
