+++ src/components/TestingPage/TestingPage.module.css
@@ -1,0 +1,81 @@
+.gridContainer {
+    display: flex;
+    flex-direction: row;
+    flex-wrap:wrap;
+    justify-content: left;
+    width: 100%;
+    border-radius: 15px;
+    overflow-y: auto;
+    padding-bottom: 10px;
+    padding-top: 10px;
+    min-height: 60vh;
+}
+
+.gridItem {
+    padding: 10px 20px 15px 20px;
+    text-align: center;
+    display: flex;
+    flex-direction: column;
+}
+
+.card {
+    width: 290px;
+    height: 190px;
+    background-color: rgb(0, 0, 0, 0.2);
+    font-size: 30px;
+    justify-content: center; 
+    transition: 0.5s;
+    align-items: center;
+    display: flex;
+    cursor: pointer;
+    border-radius: 15px;
+    box-shadow: inset 0px 0px 20px rgb(0, 0, 0, 0.5);
+    
+}
+
+.card:hover {
+    background-color: rgb(0, 0, 0, 0.3);
+}
+
+.cardSettings {
+    margin-top: 10px;
+}
+
+.cardSettingsIcon {
+    opacity: 0.4;
+    transition: 0.4s;
+}
+
+.cardSettingsIcon:hover {
+    opacity: 1.0;
+    cursor: pointer;
+}
+
+.deckManager {
+    width: 100%;
+    height: 100%;
+    display: flex;
+    flex-direction: row;
+}
+
+.deck {
+    width: 290px;
+    height: 190px;
+    background-color: rgb(255, 255, 255, 0.2);
+    justify-content: center; 
+    transition: 0.5s;
+    display: flex;
+    flex-direction: column;
+    cursor: pointer;
+    border-radius: 15px;
+    box-shadow: 0px 10px 10px rgb(0, 0, 0, 0.5);
+    padding: 1rem;
+}
+
+.deck:hover {
+    background-color: rgb(255, 255, 255, 0.15);
+}
+
+.subText {
+    opacity: 80%;
+}
