+++ src/components/TestingPage/TestingPage.tsx
@@ -1,0 +1,12 @@
+import { ReactElement } from 'react'
+
+// import styles from './TestingPage.module.css';
+
+function TestingPage(): ReactElement {
+    return ( 
+        <>
+        </>
+    );
+}
+
+export default TestingPage;
