+++ src/components/TestingPage/index.ts
@@ -1,0 +1,3 @@
+import TestingPage from './TestingPage';
+
+export default TestingPage;
